Move gzip helpers & csv parsing to separate files
